LayoutTests/fast/scrolling/mac/scrollbars/scrollbar-state-expected.txt

Test scrollbar state internals

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Document
PASS internals.verticalScrollbarState() is "enabled"
PASS internals.horizontalScrollbarState() is "none"
overflow:auto
PASS internals.verticalScrollbarState(autoScroller) is "enabled"
PASS internals.horizontalScrollbarState(autoScroller) is "none"
overflow:scroll
PASS internals.verticalScrollbarState(scrollScroller) is "enabled"
PASS internals.horizontalScrollbarState(scrollScroller) is "disabled"
PASS successfullyParsed is true

TEST COMPLETE
